MDX Refactoring
Refactored MDX components into a new MDXBlock that handles the rendering of the MDX and loading of the CodeBlock
