+++ src/components/BlogPost.js
@@ -1,15 +1,9 @@
 import * as React from 'react'
 import Layout from './layout'
-import { MDXRenderer } from 'gatsby-plugin-mdx'
-import { MDXProvider } from "@mdx-js/react"
 import { GatsbyImage, getImage } from 'gatsby-plugin-image'
 import { Link } from 'gatsby'
 import { prevLink, nextLink } from './BlogPost.module.css'
-import CodeBlock from './CodeBlock'
-
-const components = {
-  pre: CodeBlock
-}
+import { MDXBlock } from './MDXBlock'
 
 const BlogPost = ({ pageContext }) => {
   const image = getImage(pageContext.node.frontmatter.hero_image)
@@ -27,11 +21,9 @@ const BlogPost = ({ pageContext }) => {
           {pageContext.node.frontmatter.hero_image_credit_text}
         </a>
       </p>
-      <MDXProvider components={components}>
-        <MDXRenderer>
+      <MDXBlock>
           {pageContext.node.body}
-        </MDXRenderer>
-      </MDXProvider>
+      </MDXBlock>
       {pageContext.prev && <Link className={prevLink} 
         to={`/blog/${pageContext.prev.slug}`}>
           Previous: {pageContext.prev.frontmatter.title}
